+++ src/index.js
@@ -1,3 +1,6 @@
 export Styled from './styled';
-export ThemeProvider from './theme-provider';
+export ThemeProvider, {RootThemeProvider} from './theme-provider';
+export StyletronProvider from './styletron-provider';
+export LibraryProvider from './library-provider';
 export {installLibraryMeta} from './default-theme';
+export {withStyle} from './hoc/with-style';